refactor: move docs sectionhero into @repo/shared (#1570)
