+++ app/views/LearnView/learnView.tsx
@@ -1,8 +1,8 @@
 import { SectionContentCard } from "@/components/common/Card/components/SectionContentCard/sectionContentCard";
 import { StyledStack } from "@/components/Docs/components/SectionContentCards/sectionContentCards.styles";
-import { StyledSectionHero } from "@/components/Docs/components/SectionHero/sectionHero.styles";
 import { useFeatureFlag } from "@databiosphere/findable-ui/lib/hooks/useFeatureFlag/useFeatureFlag";
 import { SectionContent } from "@repo/shared/views/docs/components/SectionContent/sectionContent";
+import { StyledSectionHero } from "@repo/shared/views/docs/components/SectionHero/sectionHero.styles";
 import { Fragment, type JSX } from "react";
 import { CARDS } from "./constants";
