+++ packages/shared/views/docs/components/SectionHero/components/HeroImage/heroImage.tsx
