+++ packages/shared/views/docs/components/SectionHero/sectionHero.styles.ts
